requirements.txt
tweepy>=4.12.0
python-dotenv>=0.19.0
PyYAML>=6.0
schedule>=1.1.0
boto3>=1.28.0 